Fix heatmap points disappearing with working intensity extractor
I don't fully understand the intensity extractor concept here, but I do
know that there is no such thing as `m[2]` on the [lat, lng] points we
pass into the heatmap layer.

I've also indented the <HeatmapLayer> component in the conditional for
the brief period before Apollo loads the data.